Closes #51 update R projects chapter
Also Discussed in #34

This commit adds a discussion of creating and opening R projects to the R projects chapter. 

Because we decided to do this while in the process of writing the using git and GitHub chapter, this commit also adds a ton of new images to the 08_collaboration folder.